fix(bundleStats): Correctly handle pnpm list output (#16168)
The bundle size tool was updated to use pnpm list, instead of lerna, but
that change was incomplete because the property names in the JSON output
differs between the two tools.

I updated the command to use a shared type and unified all the `pnpm
list` code into one shared function.
