+++ .vscode/launch.json
@@ -39,6 +39,15 @@
 			"args": ["release", "-g", "client", "-v", "--no-policyCheck"],
 		},
 		{
+			"name": "flub generate bundleStats",
+			"program": "${workspaceFolder}/build-tools/packages/build-cli/bin/dev",
+			"cwd": "${workspaceFolder}",
+			"request": "launch",
+			"skipFiles": ["<node_internals>/**"],
+			"type": "node",
+			"args": ["generate", "bundleStats", "-v"],
+		},
+		{
 			// Runs the `flub release report --all` command with a debugger attached.
 			"name": "flub release report",
 			"program": "${workspaceFolder}/build-tools/packages/build-cli/bin/dev",
