+++ build-tools/packages/build-cli/src/commands/list.ts
@@ -8,16 +8,7 @@ import execa from "execa";
 import { BaseCommand } from "../base";
 import { filterPackages, parsePackageFilterFlags } from "../filter";
 import { filterFlags, releaseGroupFlag } from "../flags";
-
-/**
- * The output of `pnpm -r list` is an array of objects of this shape.
- */
-interface PnpmListEntry {
-	name: string;
-	version: string;
-	path: string;
-	private: boolean;
-}
+import { PnpmListEntry, pnpmList } from "../pnpm";
 
 /**
  * Lists all the packages in a release group in topological order.
@@ -48,16 +39,8 @@ export default class ListCommand extends BaseCommand<typeof ListCommand> {
 		}
 
 		const filterOptions = parsePackageFilterFlags(this.flags);
-		const raw = await execa(`pnpm`, [`-r`, `list`, `--depth=-1`, `--json`], {
-			cwd: releaseGroup.repoPath,
-		});
-
-		if (raw.stdout === undefined) {
-			this.error(`No output from pnpm list.`, { exit: 1 });
-		}
-
-		const parsed: PnpmListEntry[] = JSON.parse(raw.stdout);
-		const filtered = filterPackages(parsed, filterOptions)
+		const pnpmListResults = await pnpmList(releaseGroup.repoPath);
+		const filtered = filterPackages(pnpmListResults, filterOptions)
 			.reverse()
 			.map((item) => {
 				// pnpm returns absolute paths, but repo relative is more useful
